Flexible configuration support
This commit contains many changes in order to support local
configuration with increased degree of flexibility.
Session is now generic over a type that implements the BgpConfig trait.
A BasicConfig implementing BgpConfig is included.

One can implement BgpConfig for their own configs and use the
trait methods 'remote_addr_allowed' to allow incoming connections from a
prefix instead of an exact IP address. Similarly, 'remote_asn_allowed'
could be implemented in a way to allow more than one peer ASN in the
local configuration.

In order to support these concepts, some of the concepts described as
optional in RFC4271 are implemented in the FSM. This includes support
for the DelayOpen timer, and the PassiveTcpEstablishment events.

In addition to the BgpConfig trait and BasicConfig struct, the
NegotiatedConfig struct is introduced. This comprises the parameters for
a Session based on the local configuration plus the received OPEN
message. Note that this is not yet complete, e.g. capabilities are
missing.